Make sure we initialize only once.
If we have jquery < 3 it triggers the `ready`
event, and `turbolinks:load` is also triggered 
on ready. 

I was investigating a double submit error and was
in the assumption that it might be caused by a double
initialisation. It was not, but anyway this is 
cleaner. 

I removed the compatibility for old/classic turbolinks,
and use the recommended way of handling the ready
event (which should work accross all versions).